requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-jazzmin==2.6.0
django-phonenumber-field==7.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
dnspython==2.4.2
Flask==3.0.0
frozenlist==1.4.0
gunicorn
idna==3.4
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
numpy
pandas
opencv-python==4.8.1.78
Pillow==10.0.1
#psycopg2==2.9.9
psycopg2-binary==2.9.9
py-eureka-client==0.11.8
pycodestyle==2.11.0
PyMySQL==1.1.0
pytz==2023.3.post1
pyzbar==0.1.9
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
yarl==1.9.2
